Restructure project file hierarchy
