+++ flaskgallery/__init__.py
@@ -1,0 +1,5 @@
+from flask import Flask
+import os
+
+app = Flask(__name__)
+app.secret_key = os.getenv('FLASK_SECRET','OnceUponATimeThereWasAWickedWitchCalledNodeJS')
