+++ flaskgallery/static/font-awesome/css/font-awesome.css
