+++ flaskgallery/static/font-awesome/fonts/FontAwesome.otf
