+++ flaskgallery/static/font-awesome/fonts/fontawesome-webfont.ttf
