+++ flaskgallery/static/font-awesome/fonts/fontawesome-webfont.woff2
