+++ flaskgallery/static/font-awesome/less/animated.less
