+++ flaskgallery/static/font-awesome/less/bordered-pulled.less
