+++ flaskgallery/static/font-awesome/less/core.less
