+++ flaskgallery/static/font-awesome/less/fixed-width.less
