+++ flaskgallery/static/font-awesome/less/font-awesome.less
