+++ flaskgallery/static/font-awesome/less/icons.less
