+++ flaskgallery/static/font-awesome/less/list.less
