+++ flaskgallery/static/font-awesome/less/mixins.less
