+++ flaskgallery/static/font-awesome/less/path.less
