+++ flaskgallery/static/font-awesome/less/rotated-flipped.less
