+++ flaskgallery/static/font-awesome/less/stacked.less
