+++ flaskgallery/static/font-awesome/scss/_animated.scss
