+++ flaskgallery/static/font-awesome/scss/_bordered-pulled.scss
