+++ flaskgallery/static/font-awesome/scss/_core.scss
