+++ flaskgallery/static/font-awesome/scss/_fixed-width.scss
