+++ flaskgallery/static/font-awesome/scss/_icons.scss
