+++ flaskgallery/static/font-awesome/scss/_larger.scss
