+++ flaskgallery/static/font-awesome/scss/_list.scss
