+++ flaskgallery/static/font-awesome/scss/_mixins.scss
