+++ flaskgallery/static/font-awesome/scss/_path.scss
