+++ flaskgallery/static/font-awesome/scss/_rotated-flipped.scss
