+++ flaskgallery/static/font-awesome/scss/_screen-reader.scss
