+++ flaskgallery/static/font-awesome/scss/_stacked.scss
