+++ flaskgallery/static/font-awesome/scss/_variables.scss
