+++ flaskgallery/static/font-awesome/scss/font-awesome.scss
