+++ flaskgallery/templates/index.html
@@ -1,0 +1,94 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="Flask Gallery">
+    <meta name="author" content="">
+    <meta name="keywords" content="">
+    <meta name="robots" content="all">
+    <meta name="googlebot" content="all">
+
+    <title>Flask Gallery</title>
+
+    <!-- Bootstrap core CSS -->
+    <link href="{{url_for('static', filename='css/bootstrap.css')}}" rel="stylesheet">
+
+    <!-- Custom fonts for this template -->
+    <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900" rel="stylesheet">
+    <!-- Custom fonts for this template -->
+    <link rel="stylesheet" href="{{url_for('static', filename='font-awesome/css/font-awesome.min.css')}}">
+    <link rel="stylesheet" href="{{url_for('static', filename='font-awesome/css/simple-line-icons.css')}}">
+    <!-- Custom styles for this template -->
+    <link href="{{url_for('static', filename='css/flaskgallery.css')}}" rel="stylesheet">
+
+  </head>
+
+  <body id="page-top">
+  <!-- Navigation -->
+  <nav class="navbar" id="mainNav">
+    <div class="container">
+      <a class="navbar-brand" href="#page-top">Flask Gallery</a>
+    </div>
+  </nav>
+
+  <section class="default" id="gallery">
+    <div class="container">
+      <div class="section-heading text-center">
+        <h3>Example Gallery</h3>
+        <hr>
+      </div>
+    </div>
+
+    <div class="container" style="padding-top:50px;">
+      <div class="row">
+
+        {% for photo in photos %}
+        <!-- Loop through dictionary object passed from __init__.py -->
+
+        <div class="col-lg-4 col-md-6 sm-4">
+          <div class="card">
+
+            <a href="/collection/{{photo.objectID}}" class="card-view">
+            <!-- click on this image and flask will dynamically route page -->
+              <img class="card-img-top" src="{{photo.thumb}}">
+            </a>
+
+            <!-- Display attributes such as image name, year, price, etc.. -->
+            <div class="card-body">
+              <h4 class="card-title">{{photo.title}}</h4>
+              <p class="card-text">{{photo.year}}</p>
+            </div>
+
+            <!-- click on this link in the card footer and flask will dynamically route page -->
+            <div class="card-footer">
+              <a href="/collection/{{photo.objectID}}" class="card-view">View Info</a>
+            </div>
+
+          </div><!-- end card -->
+        </div><!-- end column -->
+
+        {% endfor %}
+
+      </div><!-- end row -->
+    </div><!-- end container -->
+
+  </section>
+
+    <footer>
+      <div class="container">
+        <p></p>
+      </div>
+    </footer>
+
+    <!-- Bootstrap core JavaScript -->
+    <script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
+    <script src="{{url_for('static', filename='js/bootstrap.min.js')}}"></script>
+
+    <!-- Custom scripts for this template -->
+    <script src="{{url_for('static', filename='js/new-age.min.js')}}"></script>
+
+  </body>
+
+</html>
