+++ flaskgallery/templates/object.html
@@ -1,0 +1,106 @@
+<!DOCTYPE html>
+<html lang="en">
+
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="Flask Gallery Developed by James Peters in 2018. Follow me on Twitter @JamesPeters_NZ">
+    <meta name="author" content="James Peters">
+    <meta name="keywords" content="Flask, Gallery, Photos, Photo-gallery, python, web development, design, gallery, website design">
+    <meta name="robots" content="all">
+    <meta name="googlebot" content="all">
+
+    <title>Flask Gallery</title>
+
+    <!-- Bootstrap core CSS -->
+    <link href="{{url_for('static', filename='css/bootstrap.min.css')}}" rel="stylesheet">
+
+    <!-- Custom fonts for this template -->
+    <link rel="stylesheet" href="{{url_for('static', filename='font-awesome/css/font-awesome.min.css')}}">
+    <link rel="stylesheet" href="{{url_for('static', filename='css/simple-line-icons.css')}}">
+    <link href="https://fonts.googleapis.com/css?family=Lato" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Catamaran:100,200,300,400,500,600,700,800,900" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Muli" rel="stylesheet">
+    <link href="https://fonts.googleapis.com/css?family=Alex+Brush|Homemade+Apple" rel="stylesheet">
+
+    <!-- Plugin CSS -->
+    <link rel="stylesheet" href="{{url_for('static', filename='css/device-mockups.min.css')}}">
+    <!-- Custom styles for this template -->
+    <link href="{{url_for('static', filename='css/flaskgallery.css')}}" rel="stylesheet">
+
+  </head>
+
+  <body>
+  <!-- Navigation -->
+  <nav class="navbar" id="mainNav">
+    <div class="container">
+      <a class="navbar-brand" href="/"><i class="pad-right fa fa-arrow-left"></i>Return To Gallery</a>
+    </div>
+  </nav>
+
+  <section class="default" style="min-height:95vh;">
+    <div class="container">
+      <div class="row">
+
+        <div class="col-lg-6 col-md-12 mb-4 my-auto" style="padding-bottom:60px;">
+          <div class="card">
+            <img class="card-img-top" src="{{photo.url}}" style="box-shadow: 1px 10px 40px #888888;border:17px solid white;">
+          </div>
+        </div>
+
+        <div class="col-lg-6 my-auto">
+          <div class="container-fluid">
+            <div class="row">
+              <div class="col-sm-12 my-auto">
+                <div class="section-heading text-center object-head">
+                  <h2>{{photo.title}}</h2>
+                  <hr>
+                </div>
+                <div class="feature-item object-desc text-center ">
+                  <p class="text-muted">{{photo.desc}}</p>
+                  <br>
+                  <br>
+                  <p class="text-muted">{{photo.printsize}}</p>
+                  <p class="text-muted">{{photo.framedsize}}</p>
+                  <br>
+                  <p><a href="/" class="text-muted text-center"><strong style="font-size:14px;">Back</strong></a></p>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+
+      </div><!-- end row -->
+    </div><!-- end container -->
+
+  </section>
+
+  <footer>
+    <div class="container">
+      <p>&copy; 2018 Flask Gallery by <a href="https://jamespeters.cc/">James Peters</a></p>
+    </div>
+  </footer>
+
+    <!-- Bootstrap core JavaScript -->
+    <script src="{{url_for('static', filename='js/jquery.min.js')}}"></script>
+    <script src="{{url_for('static', filename='js/popper.min.js')}}"></script>
+    <script src="{{url_for('static', filename='js/bootstrap.min.js')}}"></script>
+
+    <!-- Plugin JavaScript -->
+    <script src="{{url_for('static', filename='js/jquery.easing.min.js')}}"></script>
+
+    <!-- Custom scripts for this template -->
+    <script src="{{url_for('static', filename='js/new-age.min.js')}}"></script>
+
+    <script>
+    $(document).ready(function(){
+      function openModal(){
+      	$('.modal').modal();
+      }
+    });
+  </script>
+<script async defer src="//assets.pinterest.com/js/pinit.js"></script>
+
+  </body>
+
+</html>
