+++ flaskgallery/templates/simple-line-icons.css
