+++ startup.py
@@ -1,0 +1,1 @@
+from flaskgallery.webapp import app
